refactor (schema, ProductRepository): improve schema to accept decimal as the type of price and a limited number for the image url

+++ src/main/kotlin/ecommerce/repository/ProductRepository.kt
@@ -14,7 +14,7 @@ class ProductRepository(private val jdbcTemplate: JdbcTemplate) {
             Product(
                 rs.getLong("id"),
                 rs.getString("name"),
-                rs.getDouble("price"),
+                rs.getBigDecimal("price"),
                 rs.getString("image_url"),
             )
         }
